Hooking svgElementToPdf into jsPDF as a plug-in.
With this change, code like this becomes possible:
    var pdf = new jsPDF();
    pdf.addSVG(svgData, 25, 50, options);

In comparison, currently something along the lines of the below is required:
    var pdf = new jsPDF();
    svgElementToPdf(svgData, pdf, options);

The method name "addSVG" conflicts with https://github.com/MrRio/jsPDF/blob/master/jspdf.plugin.sillysvgrenderer.js. This isn't a problem so long as we presume that the user will only use one SVG rendering plugin at a time.